avoid some argument cast warnings on windows

+++ src/emitters/sun.cpp
@@ -179,7 +179,7 @@ public:
 		Spectrum value = 
 			m_radiance * (2 * M_PI * (1-std::cos(m_theta))) *
 			(bitmap->getWidth() * bitmap->getHeight()) 
-			/ (2 * M_PI * M_PI * nSamples);
+			/ (2.0f * M_PI * M_PI * (Float) nSamples);
 
 		for (size_t i=0; i<nSamples; ++i) {
 			Vector dir = frame.toWorld(
